requirements.txt
faiss-cpu
torch==2.3.0
numpy==1.24.4
Pillow==10.3.0
Requests==2.31.0
torchvision==0.18.0
transformers==4.40.2
flash_attn==2.5.8
streamlit
numpy
langchain_ollama